Band-aid corruption in DebEdu.mwb file
Last commit added two MESSAGE tables. One of them was incomplete
and needed to be deleted, but due to a bug in MySQL Workbench, it
could not be deleted. This commit renames the unwanted table to
MESSAGE_OLD; please ignore it whenever exporting the schema model
to SQL. You will need to change the filter for MySQL Table Objects
to exclude the MESSAGE_OLD table to do so.